Version 1.0
This is an automated home brewing system used for brewing craft beer. It
utilizes a stepper motor
to control a propane needle valve. The stepper is controlled with the
SparkFun Easy Driver board
and the related library. This valve is a very commonly used valve for
most propane burners, therefore
this code could be adapted for many other uses. The valve is controlled
based on a temperature
reading from a digital temperature sensor from Adafruit. This sensor
utilizes the OneWire library.
There is a hopper using a servo for control, this is driven with the
servo library. There is also
a propane sensor (also from SparkFun) to improve safety by measuring
propane gas levels in the
environment and force the valve shut in the case of a flame out. It is
attached with the gas breakout
board from Sparkfun. There is a rotary potentiometer on the LPG sensor
in order to set the base value
for the sensor. As of now the user interface is through the serial
monitor as well as a piezo speaker
to alert user during various stages of the brew.